BUG: Make the number of filters be an integer.
The itk::WatershedMiniPileineProgressCommand ivar m_NumberOfFilters should
be an int or unsigned int. It does not make sense for a number of filters
to be a non-integer.

Wherever a computation uses its value and should yield a float or double
ype, it should be casted.

+++ Modules/Segmentation/Watersheds/src/itkWatershedMiniPipelineProgressCommand.cxx
@@ -29,7 +29,8 @@ void WatershedMiniPipelineProgressCommand
   if ( typeid( event ) == typeid( ProgressEvent ) )
     {
     m_Filter->UpdateProgress(
-      static_cast< float >( ( m_Count + po->GetProgress() ) / m_NumberOfFilters ) );
+      static_cast< float >( ( m_Count + po->GetProgress() ) /
+      static_cast< float >( m_NumberOfFilters ) ) );
     if ( po->GetProgress() == 1.0 )
       {
       m_Count += 1.0;
@@ -47,7 +48,8 @@ void WatershedMiniPipelineProgressCommand
   if ( typeid( event ) == typeid( ProgressEvent ) )
     {
     m_Filter->UpdateProgress(
-      static_cast< float >( ( m_Count + po->GetProgress() ) / m_NumberOfFilters ) );
+      static_cast< float >( ( m_Count + po->GetProgress() ) /
+      static_cast< float >( m_NumberOfFilters ) ) );
     if ( po->GetProgress() == 1.0 )
       {
       m_Count += 1.0;
